Add files via upload
I have created a background and a working play buttong for the menu of our Game. 
I had problems with the size of the background. It was either displaying pixelated in low resolution, or it was too big and was zooming too much. 
This is the script used for making the play button.