Add files via upload
I have designed 4 different menu ideas for the game. We have decided to go for the 2nd design as it was more straightforward. If we do have more time to work on the game, I will add a score button below, that will display at least 3 previous scores by the player.